+++ src/can_moteus/can_moteus/moteus_thread_manager.py
@@ -1,0 +1,10 @@
+
+
+class MoteusThreadManager():
+
+    def __init__(self):
+        pass
+
+
+    def addMotor(self, canID, motorName):
+        pass
