+++ src/utility/canbus_mappings.py
@@ -1,0 +1,18 @@
+
+# Drivebase
+CANID_REAR_RIGHT_DRIVE_MOTOR = 20
+CANID_MID_RIGHT_DRIVE_MOTOR = 21
+CANID_FRONT_RIGHT_DRIVE_MOTOR = 22
+
+CANID_REAR_LEFT_DRIVE_MOTOR = 23
+CANID_MID_LEFT_DRIVE_MOTOR = 24
+CANID_FRONT_LEFT_DRIVE_MOTOR = 25
+
+
+# Arm
+CANID_ARM_SHOULDER_MOTOR = 1
+CANID_ARM_ELBOW_MOTOR = 2
+CANID_ARM_LEFT_WRIST_MOTOR = 3
+CANID_ARM_RIGHT_WRIST_MOTOR = 4
+CANID_ARM_TURNTABLE_MOTOR = 5
+
